Reorganize integration testing files
