--- picoCTF-web/tests/__init__.py
@@ -1,1 +1,0 @@
-"""pytest tests for the picoCTF-web components."""
