--- picoCTF-web/tests/api/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Flask API component."""
