--- picoCTF-web/tests/api/functional/__init__.py
@@ -1,1 +1,0 @@
-"""Functional tests for the API endpoints."""
