+++ picoCTF-web/tests/integration/test_users.py
@@ -1,7 +1,7 @@
 """Tests for the /api/v1/users endpoints."""
 from pytest_mongo import factories
 from pytest_redis import factories
-from ..common import ( # noqa (fixture)
+from .common import ( # noqa (fixture)
   ADMIN_DEMOGRAPHICS,
   clear_db,
   client,
